+++ ht_server/entity/sys_familytoolbox.go
@@ -1,9 +1,9 @@
 package entity
 
 type Sys_familytoolbox struct{
-	Id int
-	Familyid int
-	Toolid int
+	Id int64
+	Familyid int64
+	Toolid int64
 	//启用状态 1 or 0
 	State rune
 }
